+++ packages/component-library/src/components/Link/Link.mock.tsx
@@ -3,5 +3,6 @@ import { lorem } from 'faker';
 export default {
   variant: 'button-contained',
   href: lorem.word(),
-  text: lorem.words(2)
+  text: lorem.words(2),
+  icon: 'Instagram'
 };