+++ packages/contentful-s3-sync/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @last-rev/contentful-s3-sync
 
+## 0.1.2
+
+### Patch Changes
+
+- 7b9d6a1: Added Redis Cache loader, and cleaned up dependencies
+
 ## 0.1.1
 
 ### Patch Changes