+++ packages/contentful-sync-to-fs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @last-rev/contentful-sync-to-fs
 
+## 0.1.5
+
+### Patch Changes
+
+- 7b9d6a1: Added Redis Cache loader, and cleaned up dependencies
+- Updated dependencies [7b9d6a1]
+  - @last-rev/timer@0.1.1
+
 ## 0.1.4
 
 ### Patch Changes